requirements.txt
python-telegram-bot[webhooks]==20.8
playwright
python-dotenv
requests